moved lightgate to transfer
