+++ src/main/java/frc/robot/commands/Intake/IntakeToggle.java
@@ -4,10 +4,8 @@
 
 package frc.robot.commands.Intake;
 
-import edu.wpi.first.wpilibj.util.Color8Bit;
 import frc.robot.Constants.Intake_Constants;
 import frc.robot.RobotContainer;
-import frc.robot.subsystems.BlinkyLights;
 import frc.robot.subsystems.BlinkyLights.BlinkyLightUser;
 import frc.robot.subsystems.Intake;
 
@@ -98,21 +96,21 @@ public class IntakeToggle extends BlinkyLightUser {
         // }
     }
 
-    /*
-     * Control the blinkylights based on our Note possession.
-     *
-     * Green when we have it, red otherwise
-     */
-    @Override
-    public Color8Bit colorProvider() {
-        // make sure not is safely in our possession before going back
-        return (count >= DONE_COUNT) ? BlinkyLights.GREEN : BlinkyLights.RED;
-    };
+    // /*
+    //  * Control the blinkylights based on our Note possession.
+    //  *
+    //  * Green when we have it, red otherwise
+    //  */
+    // @Override
+    // public Color8Bit colorProvider() {
+    //     // make sure not is safely in our possession before going back
+    //     return (count >= DONE_COUNT) ? BlinkyLights.GREEN : BlinkyLights.RED;
+    // };
 
-    @Override
-    public boolean requestBlink() {
-        return false; // we want solid lights
-    }
+    // @Override
+    // public boolean requestBlink() {
+    //     return false; // we want solid lights
+    // }
 
     // Called once the command ends or is interrupted.
     @Override
@@ -125,7 +123,7 @@ public class IntakeToggle extends BlinkyLightUser {
     // Returns true when the command should end, we end when count hits DONE_COUNT
     @Override
     public boolean isFinished() {
-        count = (intake.hasNote()) ? count++ : 0; // count frames with note
+        // count = (intake.hasNote()) ? count++ : 0; // count frames with note
         // leaves in driver control
         return false;
         // return (count >= DONE_COUNT); done when we hit the count - if we want to go
